adding some more unit tests on mulmanager
